+++ script/test.sol
@@ -50,7 +50,7 @@ contract KintoDeployTestWalletScript is AASetup,KYCSignature, Script {
                 _kintoID, deployerPublicKey,
                 deployerPublicKey,
                 deployerPrivateKey, block.timestamp + 1000);
-            uint8[] memory traits = new uint8[](0);
+            uint16[] memory traits = new uint16[](0);
             _kintoID.mintIndividualKyc(sigdata, traits);
         }
