+++ src/interfaces/IKintoID.sol
@@ -13,7 +13,7 @@ interface IKintoID {
         uint8 sanctionsCount;
         bool individual;
         BitMapsUpgradeable.BitMap traits;
-        BitMapsUpgradeable.BitMap sanctions;
+        BitMapsUpgradeable.BitMap sanctions; // Follows ISO-3661 numeric codes https://en.wikipedia.org/wiki/ISO_3166-1_numeric
     }
 
     struct SignatureData {
@@ -27,27 +27,27 @@ interface IKintoID {
     struct MonitorUpdateData {
         bool isTrait; // otherwise sanction
         bool isSet; // otherwise remove
-        uint8 index;
+        uint16 index;
     }
 
     /* ============ State Change ============ */
 
 
-    function mintIndividualKyc(SignatureData calldata _signatureData, uint8[] memory _traits) external;
+    function mintIndividualKyc(SignatureData calldata _signatureData, uint16[] memory _traits) external;
 
-    function mintCompanyKyc(SignatureData calldata _signatureData, uint8[] memory _traits) external;
+    function mintCompanyKyc(SignatureData calldata _signatureData, uint16[] memory _traits) external;
 
     function burnKYC(SignatureData calldata _signatureData) external;
 
     function setURI(string memory newuri) external;
 
-    function addTrait(address _account, uint8 _traitId) external;
+    function addTrait(address _account, uint16 _traitId) external;
 
-    function removeTrait(address _account, uint8 _traitId) external;
+    function removeTrait(address _account, uint16 _traitId) external;
 
-    function addSanction(address _account, uint8 _countryId) external;
+    function addSanction(address _account, uint16 _countryId) external;
 
-    function removeSanction(address _account, uint8 _countryId) external;
+    function removeSanction(address _account, uint16 _countryId) external;
 
     function monitor(address[] memory _accounts, MonitorUpdateData[][] memory _traitsAndSanctions) external;
 
@@ -63,7 +63,7 @@ interface IKintoID {
 
     function isSanctionsSafe(address _account) external view returns (bool);
 
-    function isSanctionsSafeIn(address _account, uint8 _countryId) external view returns (bool);
+    function isSanctionsSafeIn(address _account, uint16 _countryId) external view returns (bool);
 
     function isCompany(address _account) external view returns (bool);
 
@@ -71,7 +71,7 @@ interface IKintoID {
 
     function mintedAt(address _account) external view returns (uint256);
 
-    function hasTrait(address _account, uint8 index) external view returns (bool);
+    function hasTrait(address _account, uint16 index) external view returns (bool);
 
     function traits(address _account) external view returns (bool[] memory);
