+++ test/KintoID.t.sol
@@ -98,7 +98,7 @@ contract KintoIDTest is KYCSignature {
 
     function testMintIndividualKYC() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](0);
+        uint16[] memory traits = new uint16[](0);
         vm.startPrank(_kycProvider);
         assertEq(_kintoIDv1.isKYC(_user), false);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
@@ -112,7 +112,7 @@ contract KintoIDTest is KYCSignature {
 
     function testMintCompanyKYC() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](2);
+        uint16[] memory traits = new uint16[](2);
         traits[0] = 2;
         traits[1] = 5;
         vm.startPrank(_kycProvider);
@@ -128,7 +128,7 @@ contract KintoIDTest is KYCSignature {
 
     function testMintIndividualKYCWithInvalidSender() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_user);
         vm.expectRevert('Invalid Provider');
@@ -137,7 +137,7 @@ contract KintoIDTest is KYCSignature {
 
     function testMintIndividualKYCWithInvalidSigner() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 5, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         vm.expectRevert('Invalid Signer');
@@ -146,7 +146,7 @@ contract KintoIDTest is KYCSignature {
 
     function testMintIndividualKYCWithInvalidNonce() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
@@ -156,7 +156,7 @@ contract KintoIDTest is KYCSignature {
 
     function testMintIndividualKYCWithExpiredSignature() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp - 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         vm.expectRevert('Signature has expired');
@@ -167,7 +167,7 @@ contract KintoIDTest is KYCSignature {
 
     function testBurnKYC() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
@@ -179,7 +179,7 @@ contract KintoIDTest is KYCSignature {
 
     function testOnlyProviderCanBurnKYC() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
@@ -199,7 +199,7 @@ contract KintoIDTest is KYCSignature {
 
     function testBurningTwiceFails() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
@@ -235,7 +235,7 @@ contract KintoIDTest is KYCSignature {
     function testSettingTraitsAndSanctions() public {
         vm.startPrank(_kycProvider);
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         address[] memory accounts = new address[](1);
@@ -315,7 +315,7 @@ contract KintoIDTest is KYCSignature {
 
     function testFailTransfersAreDisabled() public {
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(_kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](1);
+        uint16[] memory traits = new uint16[](1);
         traits[0] = 1;
         vm.startPrank(_kycProvider);
         assertEq(_kintoIDv1.isKYC(_user), false);
