+++ test/KintoWallet.t.sol
@@ -106,7 +106,7 @@ contract KintoWalletTest is UserOp, KYCSignature {
         // Mint an nft to the owner
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(
             _kintoIDv1, _owner, _owner, 1, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](0);
+        uint16[] memory traits = new uint16[](0);
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         vm.stopPrank();
@@ -557,7 +557,7 @@ contract KintoWalletTest is UserOp, KYCSignature {
         // Mint NFT to new owner and burn old
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(
             _kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](0);
+        uint16[] memory traits = new uint16[](0);
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         sigdata = _auxCreateSignature(_kintoIDv1, _owner, _owner, 1, block.timestamp + 1000);
@@ -603,7 +603,7 @@ contract KintoWalletTest is UserOp, KYCSignature {
         // Mint NFT to new owner
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(
             _kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](0);
+        uint16[] memory traits = new uint16[](0);
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         vm.stopPrank();
@@ -661,7 +661,7 @@ contract KintoWalletTest is UserOp, KYCSignature {
         // Mint NFT to new owner and burn old
         IKintoID.SignatureData memory sigdata = _auxCreateSignature(
             _kintoIDv1, _user, _user, 3, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](0);
+        uint16[] memory traits = new uint16[](0);
         vm.startPrank(_kycProvider);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
         sigdata = _auxCreateSignature(_kintoIDv1, _owner, _owner, 1, block.timestamp + 1000);
